move stuff where it should be, default to xlsx
