+++ src/scss/_spreadsheet-editor.scss
@@ -1,0 +1,20 @@
+/* spreadsheet editor iframe*/
+#spreadsheetEditorRoot {
+  background: $mainbackground;
+  border: 0;
+  height: 100%;
+  inset: 0;
+  position: absolute;
+  width: 100vw;
+}
+
+#spreadsheetEditorDiv {
+  height: 40vh;
+  overflow-y: hidden;
+
+  iframe {
+    border: 0;
+    height: 100%;
+    width: 100%;
+  }
+}
